requirements.txt
pycti==6.6.7
newspaper3k>=0.2.8
requests>=2.25.1
beautifulsoup4>=4.11.1
pdfkit>=1.0.0
PyYAML>=6.0
lxml==4.9.2
lxml-html-clean>=0.1.0